fix(tests): split up conftest.py into several different test utils files
