+++ tests/unit/api/v1/test_ranges.py
@@ -1,6 +1,5 @@
 import copy
 import random
-import string
 from unittest.mock import AsyncMock
 
 import pytest
@@ -18,7 +17,7 @@ from src.app.schemas.range_schemas import (
     DeployedRangeSchema,
 )
 from src.app.schemas.secret_schema import SecretSchema
-from tests.conftest import authenticate_client
+from tests.api_test_utils import authenticate_client
 
 from .config import (
     BASE_ROUTE,
@@ -29,71 +28,6 @@ from .config import (
 )
 
 pytestmark = pytest.mark.unit
-
-
-###### Test /ranges/deploy #######
-
-
-# async def test_deploy_without_enc_key(client: AsyncClient) -> None:
-#     """Test that attempting to deploy a range without being logged in will fail since the encryption key was not given from successful login."""
-#     assert await authenticate_client(client), "Failed to authenticate to API"
-
-#     for cookie in client.cookies.jar:
-#         if cookie.name == "enc_key":
-#             cookie.value = ""
-#             break
-
-#     response = await client.post(
-#         f"{BASE_ROUTE}/ranges/deploy", json=valid_range_deploy_payload
-#     )
-#     assert response.status_code == status.HTTP_401_UNAUTHORIZED
-
-
-# async def test_deploy_without_valid_enc_key(auth_client: AsyncClient) -> None:
-#     """Test that attempting to deploy a range with an invalid encryption key will fail."""
-#     modified_enc_key = "in*vali*^%$"
-#     auth_client.cookies.update({"enc_key": modified_enc_key})
-#     response = await auth_client.post(
-#         f"{BASE_ROUTE}/ranges/deploy",
-#         json=valid_range_deploy_payload,
-#     )
-#     assert response.status_code == status.HTTP_400_BAD_REQUEST
-
-
-# async def test_deploy_without_valid_range_blueprint(auth_client: AsyncClient) -> None:
-#     """Test that attempting to deploy a range with a non-existent range blueprint will fail."""
-#     enc_key = "VGhpcyBpcyBhIHRlc3Qgc3RyaW5nIGZvciBiYXNlNjQgZW5jb2Rpbmcu"
-#     auth_client.cookies.update({"enc_key": enc_key})
-#     non_existent_range_deploy_payload = copy.deepcopy(valid_range_deploy_payload)
-#     non_existent_range_deploy_payload["blueprint_id"] = random.randint(  # noqa: S311
-#         -666, -69
-#     )
-#     response = await auth_client.post(
-#         f"{BASE_ROUTE}/ranges/deploy",
-#         json=valid_range_deploy_payload,
-#     )
-#     assert response.status_code == status.HTTP_404_NOT_FOUND
-
-
-# async def test_deploy_without_valid_private_key(auth_client: AsyncClient) -> None:
-#     """Test that attempting to deploy a range without valid private key will fail."""
-#     enc_key = "VGhpcyBpcyBhIHRlc3Qgc3RyaW5nIGZvciBiYXNlNjQgZW5jb2Rpbmcu"
-#     auth_client.cookies.update({"enc_key": enc_key})
-#     response = await auth_client.post(
-#         f"{BASE_ROUTE}/blueprints/ranges",
-#         json=valid_blueprint_range_create_payload,
-#     )
-#     assert response.status_code == status.HTTP_200_OK
-#     blueprint_id = int(response.json()["id"])
-
-#     blueprint_deploy_payload = copy.deepcopy(valid_range_deploy_payload)
-#     blueprint_deploy_payload["blueprint_id"] = blueprint_id
-
-#     response = await auth_client.post(
-#         f"{BASE_ROUTE}/ranges/deploy",
-#         json=blueprint_deploy_payload,
-#     )
-#     assert response.status_code == status.HTTP_400_BAD_REQUEST
 
 
 async def test_deploy_without_valid_secrets(
@@ -222,44 +156,6 @@ async def test_deploy_range_deploy_success(
     assert int(response.json()["id"])
 
 
-###### Test /ranges/destroy #######
-
-
-# async def test_destroy_with_invalid_id(auth_client: AsyncClient) -> None:
-#     """Test that attempting to destroy a range without a valid non-int ID with fail."""
-#     random_str = "".join(
-#         random.choice(string.ascii_letters)  # noqa: S311
-#         for i in range(random.randint(1, 10))  # noqa: S311
-#     )
-#     response = await auth_client.delete(f"{BASE_ROUTE}/ranges/{random_str}")
-#     assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
-
-
-# async def test_destroy_without_enc_key(client: AsyncClient) -> None:
-#     """Test that attempting to destroy a range without being logged in will fail since the encryption key was not given from successful login."""
-#     assert await authenticate_client(client), "Failed to authenticate to API"
-
-#     for cookie in client.cookies.jar:
-#         if cookie.name == "enc_key":
-#             cookie.value = ""
-#             break
-
-#     response = await client.delete(
-#         f"{BASE_ROUTE}/ranges/{random.randint(1, 100)}"  # noqa: S311
-#     )
-#     assert response.status_code == status.HTTP_401_UNAUTHORIZED
-
-
-# async def test_destroy_without_valid_enc_key(auth_client: AsyncClient) -> None:
-#     """Test that attempting to destroy a range with an invalid encryption key will fail."""
-#     modified_enc_key = "in*vali*^%$"
-#     auth_client.cookies.update({"enc_key": modified_enc_key})
-#     response = await auth_client.delete(
-#         f"{BASE_ROUTE}/ranges/{random.randint(1, 100)}"  # noqa: S311
-#     )
-#     assert response.status_code == status.HTTP_400_BAD_REQUEST
-
-
 async def test_destroy_without_valid_range_owner(
     auth_client: AsyncClient,
     client: AsyncClient,
@@ -294,16 +190,6 @@ async def test_destroy_without_valid_range_owner(
         pytest.fail("Failed to authenticate client to API!")
     response = await client.delete(f"{BASE_ROUTE}/ranges/{user1_range_id}")
     assert response.status_code == status.HTTP_404_NOT_FOUND
-
-
-# async def test_destroy_without_valid_range(auth_client: AsyncClient) -> None:
-#     """Test that attempting to destroy a non-existent range will fail."""
-#     enc_key = "VGhpcyBpcyBhIHRlc3Qgc3RyaW5nIGZvciBiYXNlNjQgZW5jb2Rpbmcu"
-#     auth_client.cookies.update({"enc_key": enc_key})
-#     response = await auth_client.delete(
-#         f"{BASE_ROUTE}/ranges/{random.randint(-420, -69)}"  # noqa: S311
-#     )
-#     assert response.status_code == status.HTTP_404_NOT_FOUND
 
 
 async def test_destroy_decrypt_secrets_failure(
@@ -519,15 +405,6 @@ async def test_destroy_range_destroy_success(
     assert response.status_code == status.HTTP_200_OK
 
 
-# async def test_get_range_headers_empty_list(client: AsyncClient) -> None:
-#     """Test that we get a 404 when there are no deployed range headers."""
-#     # Make a new user that will not have deployed ranges
-#     assert await authenticate_client(client)
-
-#     response = await client.get(f"{BASE_ROUTE}/ranges")
-#     assert response.status_code == status.HTTP_404_NOT_FOUND
-
-
 async def test_get_range_headers_success(
     auth_client: AsyncClient, monkeypatch: pytest.MonkeyPatch
 ) -> None:
@@ -546,13 +423,6 @@ async def test_get_range_headers_success(
     assert response_data == [header.model_dump(mode="json")]
 
 
-# async def test_get_range_details_non_existent(auth_client: AsyncClient) -> None:
-#     """Test that we get a 404 when we request a non-existent range."""
-#     non_existent_id = random.randint(-420, -69)  # noqa: S311
-#     response = await auth_client.get(f"{BASE_ROUTE}/ranges/{non_existent_id}")
-#     assert response.status_code == status.HTTP_404_NOT_FOUND
-
-
 async def test_get_range_details_success(
     auth_client: AsyncClient, monkeypatch: pytest.MonkeyPatch
 ) -> None:
@@ -569,13 +439,6 @@ async def test_get_range_details_success(
 
     response_data = response.json()
     assert response_data == test_range.model_dump(mode="json")
-
-
-# async def test_get_range_key_non_existent(auth_client: AsyncClient) -> None:
-#     """Test that we get a 404 when we request a non-existent range's key."""
-#     non_existent_id = random.randint(-420, -69)  # noqa: S311
-#     response = await auth_client.get(f"{BASE_ROUTE}/ranges/{non_existent_id}/key")
-#     assert response.status_code == status.HTTP_404_NOT_FOUND
 
 
 async def test_get_range_key_success(
